requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
gradio==4.36.1
huggingface-hub==0.23.3
ipykernel==6.29.4
ipython==8.25.0
joblib==1.4.2
keras==3.3.3
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==1.26.4
opencv-python==4.10.0.82
pandas==2.2.2
pillow==10.3.0
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
streamlit==1.35.0
sympy==1.12.1
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.4
ultralytics==8.2.29
ultralytics-thop==0.2.7